fix: :bug: Fixes the models, migrations and seeders for Sales and Products tables

+++ back-end/src/database/migrations/20220920194544-create-products.js
@@ -13,7 +13,7 @@ module.exports = {
         allowNull: false,
       },
       price: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(10,2),
         allowNull: false,
       },
       urlImage: {
