+++ back-end/src/database/migrations/20220920195207-create-sales.js
@@ -21,7 +21,7 @@ module.exports = {
         references: { model: 'users', key: 'id' },
       },
       totalPrice: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(10,2),
         field: 'total_price',
         allowNull: false,
       },
